separate profile into own module, add board tests
